Total refactor of comments management and fetching
To accommodate the unthreaded structure of /api/morecomments, we reimplement the backing data structure of the comments view to maintain a tree. Whenever we need to add or remove comments or More placeholders from the view,
we perform this operation on the tree, and then flatten the tree into an array in the appropriate order using a pre-order traversal.