Reimplement PreferencesData as a static singleton
Passing this an environmentObject would be ideal, but it is unwieldily to do that between windows, and often results an a crash even when switching between children of the root view.
Either I am not implementing this correctly, it is not meant for what I am using it for, or it is brittle.