Reimplement `heightResizable` with Preferences instead of NSView background hack
I could not get this to work initially, but after changing the preference key to pass the frame instead of just the bounds, it started working.
This is much more efficient and smooth than the NSView hack, it resizes in realtime, results in no runtime crashes. and works with WebImage (and should work with any view)